Class/CB cleanup/fixes
* Set primary key identities (hopefully)
* Moved entity classes for entity framework into own folder
