--- ServerProgram/IClientListener.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Threading;
-
-namespace ServerProgram
-{
-    interface IClientListener
-    {
-        Task BeginWaitForConnections();
-        void EndWaitForConnections();
-        void Disconnect();
-    }
-}
