+++ ServerProgram/ServerProgram.csproj
@@ -52,27 +52,25 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Account.cs" />
-    <Compile Include="Building.cs" />
+    <Compile Include="Entities\Account.cs" />
+    <Compile Include="Entities\Building.cs" />
     <Compile Include="CampusContext.cs" />
-    <Compile Include="CampusModel.cs" />
     <Compile Include="ClientHandler.cs" />
     <Compile Include="ClientListener.cs" />
-    <Compile Include="DormRoom.cs" />
-    <Compile Include="Floor.cs" />
-    <Compile Include="IClientHandler.cs" />
-    <Compile Include="IClientListener.cs" />
+    <Compile Include="Entities\DormRoom.cs" />
+    <Compile Include="Entities\Floor.cs" />
     <Compile Include="Mappers\AccountMapper.cs" />
     <Compile Include="Mappers\StudentMapper.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Reservation.cs" />
-    <Compile Include="Student.cs" />
+    <Compile Include="Entities\Reservation.cs" />
+    <Compile Include="Entities\Student.cs" />
     <Compile Include="Test\Test.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
